fix: properly configure postcss.config.js for refactored theme
After the project was refactored to put the theme in a separate directory,
it broke the whitelister for the `purgecss` module. It seems the Node
environment in which Hugo Pipes runs doesn't operate on Hugo's merged
filesystem. This was resolved by adding globs that check for the relevant
directories under `./themes/*/` as well as under `./`.

Signed-off-by: Nick Anderegg <nick@anderegg.io>