feat: implement menus to be arranged totally on the fly
Previously, menu generation required a `docs` menu config in the menu.yaml
configuration file. Now, the documentation navigation menu is generated
totally based on the filesystem structure as well as the weights
assigned to each page/section.

Additionally, there was a slight refactoring to make all of the sections
their own top-level sections.

Signed-off-by: Nick Anderegg <nick@anderegg.io>