refactor: fix section weights and structuring
This commit primarily restructures the navigation of the various
sections of the site. In addition, in several files, unneeded frontmatter
metadata fields are removed, as are manually-written section contents
in _index.md files. These manual contents are removed because section
pages will automatically generate a contents list for the section.

Signed-off-by: Nick Anderegg <nick.anderegg@zymbit.com>