refactor: define layout for pages in config file
Because this is primarily a docs site, we want all of our pages to use the
docs layout. Normally, the layout is chosen by the name of the first
directory under `content/`. To avoid needing to put `type: docs`
in every page we can define cascading frontmatter in the site config
that causes every page below this to inherit `type: docs`

Any other key-value pairs added under the `cascade` key will be inherited by
every other site page. More information about front matter cascades:
https://gohugo.io/content-management/front-matter#front-matter-cascade

Additionally, setting an alternative type in a page will not affect the type
configured for the nodes below--unless the cascade key is specified
in the frontmatter of that page.

Signed-off-by: Nick Anderegg <nick.anderegg@zymbit.com>